requirements.txt
haystack-ai
qdrant-client
qdrant-haystack
fastembed-haystack
requests
python-dotenv